# lab1/docs/swagger.yaml
basePath: /
definitions:
...
  ds.DeleteMedMmPvlcCalculationRequest:
    properties:
      card_id:
        type: integer
      pvlc_med_formula_id:
        type: integer
    required:
    - card_id
    - pvlc_med_formula_id
    type: object
  ds.LoginMedUserRequest:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
...
  ds.MedUserRegistrationRequest:
    properties:
      is_moderator:
        type: boolean
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  ds.MedUserResponse:
    properties:
      id:
        type: integer
      is_moderator:
        type: boolean
      login:
        type: string
    type: object
...
  ds.UpdateMedMmPvlcCalculationAPIRequest:
    properties:
      card_id:
        type: integer
      data:
        $ref: '#/definitions/ds.UpdateMedMmPvlcCalculationRequest'
      pvlc_med_formula_id:
        type: integer
    required:
    - card_id
    - data
    - pvlc_med_formula_id
    type: object
  ds.UpdateMedMmPvlcCalculationRequest:
    properties:
      input_height:
        type: number
    required:
    - input_height
    type: object
  ds.UpdateMedUserRequest:
    properties:
      password:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@lungcalc.ru
    name: API Support
    url: http://localhost:8080
  description: |-
    API для расчета должной жизненной емкости легких (ДЖЕЛ)
    Лабораторная работа 4 - Добавление аутентификации и авторизации
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Lung Capacity Calculation API
  version: "1.0"
paths:
...
  /api/med-mm-pvlc-calculations:
    delete:
      consumes:
      - application/json
      description: Удаляет связь формулы с заявкой (только для владельца черновика)
      parameters:
      - description: Данные для удаления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.DeleteMedMmPvlcCalculationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление расчета из заявки
      tags:
      - med_calculations
    put:
      consumes:
      - application/json
      description: Обновляет данные расчета в заявке (ввод роста)
      parameters:
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.UpdateMedMmPvlcCalculationAPIRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновление расчета в заявке
      tags:
      - med_calculations
  /api/med-users/login:
    post:
      consumes:
      - application/json
      description: Выполняет вход пользователя
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.LoginMedUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Аутентификация пользователя (старый endpoint)
      tags:
      - med_users
...
  /api/med-users/profile:
    get:
      description: Возвращает информацию о пользователе по ID (для демонстрации)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.MedUserResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получение профиля пользователя (старый endpoint)
      tags:
      - med_users
    put:
      consumes:
      - application/json
      description: Обновляет данные профиля пользователя
      parameters:
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.UpdateMedUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновление профиля пользователя
      tags:
      - med_users
  /api/med-users/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя в системе
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ds.MedUserRegistrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Регистрация нового пользователя
      tags:
      - med_users
...
securityDefinitions:
  BearerAuth:
    description: 'JWT токен для аутентификации. Формат: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"